+++ frontend/src/components/Footer.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Footer = () => {
+  return (
+    <footer style={{ padding: '10px', backgroundColor: '#f1f1f1', textAlign: 'center' }}>
+      <p>© 2024 Your Company. All rights reserved.</p>
+    </footer>
+  );
+};
+
+export default Footer;
